EN-2311 Abstract job executors, refactor code accordingly, adds commands for cms components, update tests
